fix of the data type mis-matched problem

+++ src/assembly/mzPack/GCMSConvertor.vb
@@ -90,10 +90,10 @@ Public Class GCMSnetCDF
 
     Public Iterator Function readMzmatrix() As IEnumerable(Of Double())
         Dim offset As Integer = 0
-        Dim v As Short()
+        Dim v As Single()
 
         For Each width As Integer In point_count
-            v = New Short(width - 1) {}
+            v = New Single(width - 1) {}
             Array.ConstrainedCopy(mass_values, offset, v, Scan0, width)
             Yield v.Select(Function(si) CDbl(si)).ToArray
             offset += width
